storage/zfs: move some helper functions to storage_zfs_utils.go
